Patches with checkin+ should not prevent a bug from being in "To Fix"
Prior to this change, bugs with a reviewed but checked in patch did not
appear on any of the Bugzilla-Todos sections, since they were also
excluded from "To Check In". Instead now they appear on "To Fix", since
there is either (a) work still to be done in the bug, or (b) the user
forgot to close the bug, in which case at least showing it _somewhere_
will mean they are reminded to close it.

Fixes #69.