Remove Google Analytics
The GA code was added in January 2013, so is presumably still using
Heather's account. Now that the project has been passed on, I'm not sure
it makes sense to continue reporting to her GA account. In addition,
there are some Mozillians who avoid sites that use GA, so let's just
remove it.